dnn_mnist/pyproject.toml
[project]
name = "dnn-mnist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-contrib-python>=4.11.0.86",
    "pillow>=12.0.0",
    "tensorflow>=2.20.0",
]
